Move client id into its own package
